do not cd to non-existent directory.
git-svn-id: https://ice.txcorp.com/svnrepos/code/bilder/trunk@2563 a9e16012-fd73-49ee-95cb-036de7809213

+++ packages/oce_aux.sh
@@ -17,9 +17,11 @@
 
 setOceTriggerVars() {
 # Remove old repo
-  newocerem=`(cd $PROJECT_DIR/oce; git remote -v | grep Tech-XCorp)`
-  if test -z "$newocerem"; then
-    rm -rf $PROJECT_DIR/oce
+  if test -d $PROJECT_DIR/oce; then
+    newocerem=`(cd $PROJECT_DIR/oce; git remote -v | grep Tech-XCorp)`
+    if test -z "$newocerem"; then
+      rm -rf $PROJECT_DIR/oce
+    fi
   fi
 # Standard vars
   OCE_REPO_URL=https://github.com/Tech-XCorp/oce.git
